deep link & debug auth
