+++ app/config/local.yml
@@ -1,8 +1,7 @@
 parameters:
-  security_type: 'ldap' # debug, ntlm, ldap
+  security_type: 'demo' # demo, ntlm, ldap
   security_ldap_url: 'ldap.forumsys.com'
   security_ldap_dn_string: 'cn=read-only-admin,dc=example,dc=com' # {username}
   security_ntlm_domain: 'workspace'
-  security_debug_username: 'mustermann'
 
   notification_email_sender: ~
